scrape/requirements/request.txt
GET /DN/Audit/DegreeAudit.aspx?pageid=audit&degreeID=150 HTTP/1.1
Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7
Accept-Encoding: gzip, deflate, br, zstd
Accept-Language: en-US,en;q=0.9
Connection: keep-alive
Cookie: fpestid=cvScEwMst7cLU-7Efzqra0ZoXRas7uIQIUFmC9WGs9MOLNLV55HK74eRQ6L8i0gOdJlMsw; BIGipServerdn_os_ver_12-http-Pool=737090988.20480.0000; kndctr_1B6E34B85282A0AC0A490D44_AdobeOrg_identity=CiY3NzgxOTA4OTExNjIwMjU1ODQ5NDU3MTUwMTgxOTM2Mjc3MTUxOFIQCNW56cChMhgBKgNWQTYwAfAB1bnpwKEy; lastRskxRun=1727025823736; _hjSessionUser_2593350=eyJpZCI6ImFhYzA4OGU5LWUyMzEtNThlYS05YjM1LTg2YTg5YjY1YjJlMCIsImNyZWF0ZWQiOjE3MjgzMzMzMzUyNjQsImV4aXN0aW5nIjp0cnVlfQ==; AMCVS_CB68E4BA55144CAA0A4C98A5%40AdobeOrg=1; AMCV_CB68E4BA55144CAA0A4C98A5%40AdobeOrg=1585540135%7CMCIDTS%7C20010%7CMCMID%7C74131072186103455623634261956361102349%7CMCAAMLH-1729391626%7C7%7CMCAAMB-1729391626%7CRKhpRz8krg2tLO6pguXWp5olkAcUniQYPHaMWWgdJ3xzPWQmdj0y%7CMCOPTOUT-1728794026s%7CNONE%7CMCAID%7CNONE%7CvVersion%7C4.4.0; s_cc=true; djm="id=902adf03-7569-429c-b246-75ce7fa50d40"; s_sq=%5B%5BB%5D%5D; _cc_id=5f8a0b03eb0f670b515973dc10dd8acf; amplitude_id_9f6c0bb8b82021496164c672a7dc98d6_edmrutgers.edu=eyJkZXZpY2VJZCI6ImZhZTliNGExLWQ4OTktNGMyNy04MzA1LTY3NWY5N2RjM2ZjZlIiLCJ1c2VySWQiOm51bGwsIm9wdE91dCI6ZmFsc2UsInNlc3Npb25JZCI6MTcyODc5Mzc5ODg5MSwibGFzdEV2ZW50VGltZSI6MTcyODc5Mzg0MDkyOSwiZXZlbnRJZCI6MCwiaWRlbnRpZnlJZCI6MTAsInNlcXVlbmNlTnVtYmVyIjoxMH0=; amplitude_id_408774472b1245a7df5814f20e7484d0rutgers.edu=eyJkZXZpY2VJZCI6ImZiZTQzZGYwLWRiY2QtNDk3Yy05ZDYwLWFiNWQxZDVmMDljMyIsInVzZXJJZCI6Ii0xNzEyOTc0OSIsIm9wdE91dCI6ZmFsc2UsInNlc3Npb25JZCI6MTcyODc5Mzc3Nzc1NSwibGFzdEV2ZW50VGltZSI6MTcyODc5Mzg0MDk3MywiZXZlbnRJZCI6NDAsImlkZW50aWZ5SWQiOjU2LCJzZXF1ZW5jZU51bWJlciI6OTZ9; _ga_VR37SQR64F=GS1.1.1728794636.2.0.1728794637.59.0.0; _ga_HW17LFESG1=GS1.1.1728929503.2.0.1728929503.0.0.0; _ga_H7QH4QNXJN=GS1.1.1728929503.2.0.1728929503.0.0.0; _ga_CN4D4Q9RS7=GS1.1.1729178084.1.0.1729178094.0.0.0; _ga_ZP4HK22193=GS1.1.1729361771.1.1.1729361787.0.0.0; _ga_MH9FSG2KVN=GS1.1.1729371702.2.0.1729371715.0.0.0; _hjSessionUser_3158812=eyJpZCI6IjlmMTY5ZDU4LTk1YTktNWY5OS05MmZjLTlhODc0ZjliMDRhMSIsImNyZWF0ZWQiOjE3MjkzOTQ2MzEzNzUsImV4aXN0aW5nIjp0cnVlfQ==; _ga_NR5RPD5PCZ=GS1.1.1729461969.3.0.1729462068.0.0.0; _ga_8KMWEC10BB=GS1.1.1729554176.3.0.1729554178.0.0.0; _ga_J1DSRJRQZ1=GS1.1.1729710513.1.1.1729714190.0.0.0; _ga_Q9231LEHXD=GS1.1.1729712713.1.1.1729714192.0.0.0; _ga_H4R35LHY9K=GS1.1.1730385607.1.1.1730385835.0.0.0; sascas_4_envelope_3=038715706383741730750044; _ga_29NEBSX93J=GS1.1.1730824238.1.1.1730824405.0.0.0; _fbp=fb.1.1731079092149.3950818263243147; twk_uuid_5b4dfb5edf040c3e9e0ba5fc=%7B%22uuid%22%3A%221.70hsskrXAauZUMBYgfSb5Dg3TqiwZk53VU1kJctWsTIpNLIwteaV5rZ6jSdgcJMWtya6lgVpEBSlXU4KUIdXQFLWQ9HqLdyWUwvaJ9c9XiYIU91GaZvU%22%2C%22version%22%3A3%2C%22domain%22%3A%22rutgers.edu%22%2C%22ts%22%3A1731081458051%7D; _ga_ZYY85J4EN8=GS1.1.1731298592.1.1.1731298650.0.0.0; _ga_R6HEQ3RYNP=GS1.1.1731441273.6.1.1731441897.0.0.0; _ga_D9FXXT73YN=GS1.1.1732403087.1.1.1732403537.0.0.0; _hjSessionUser_5061430=eyJpZCI6IjUxZDVlNWNjLWYyODEtNWViNC1iZDgxLWFiMDk2OTg5YjVkNSIsImNyZWF0ZWQiOjE3MzQ4MjM4OTU1MDAsImV4aXN0aW5nIjpmYWxzZX0=; _ga_JDZ3MW6CDW=GS1.2.1734823895.1.0.1734823895.60.0.0; _ga_6VMLQ8DVK9=GS1.1.1734823894.1.1.1734824111.0.0.0; sascas_4_envelope_38=091996502879541735917571; _ga_F78K85ZPMG=GS1.1.1735930752.5.0.1735931406.0.0.0; _ga_JCHSMVR2MB=GS1.2.1735931570.1.0.1735931570.0.0.0; _ga_RQCVQRJLEX=GS1.2.1736475362.1.1.1736475383.0.0.0; _ga_TCRF23SX9L=GS1.1.1736475398.1.0.1736475401.0.0.0; BCSessionID=d2016f6b-e009-4e55-acff-d8b004ee730a; _gcl_au=1.1.1266068347.1736879599; _ga_B6BCKQE1XW=GS1.1.1737422035.4.0.1737422035.0.0.0; _ga_BMN5F2WXL0G-BMN5F2WXL0G-BMN5F2WXL0=GS1.1.1737559779.6.0.1737559779.0.0.0; _ga_BMN5F2WXL0=GS1.1.1737559779.9.0.1737560009.60.0.0; _ga_3K77D0NWXZ=GS1.1.1737647065.1.1.1737647084.0.0.0; _ga_RX0G5XBWJW=GS1.1.1737647065.1.1.1737647085.0.0.0; EssUserTrk=b2ad6391.62c682087ba51; _ga_2L8851K40M=GS1.1.1738078241.12.0.1738078241.0.0.0; _ga_YGLWDXH2GH=GS1.1.1738177031.14.0.1738177031.60.0.0; _ga_RWCX59EEM6=GS1.1.1738177031.14.0.1738177031.0.0.0; _ga_M6FS8HG1PG=GS1.1.1738462912.72.0.1738462912.0.0.0; _ga_4BJDT7REEF=GS1.1.1738860024.20.0.1738860025.0.0.0; _ga_C9TCE1C461=GS1.1.1738865893.21.0.1738865893.60.0.0; _ga_T2SXJM5L7G=GS1.1.1738904527.10.0.1738904532.0.0.0; _ga_VC05CXJG7L=GS1.1.1738904613.1.0.1738906296.0.0.0; _ga_0SP8PB5CPS=GS1.1.1738950091.7.1.1738950103.48.0.0; ASP.NET_SessionId=pa5wzulzbdeh4vogs0l4avtp; _ga_LGEQN89JC7=GS1.1.1738956858.1.0.1738956879.0.0.0; _ga_DVCN4VQT8D=GS1.1.1739301283.7.0.1739301290.0.0.0; _ga_LTZD2MY0RB=GS1.1.1739401703.2.0.1739401703.0.0.0; _ga_KQ936KPQQ8=GS1.1.1739461303.20.0.1739461303.0.0.0; _clck=1r7v2ba%7C2%7Cftf%7C0%7C1566; _ga_NQ7HLJ30E3=GS1.1.1739567769.2.1.1739567791.0.0.0; _ga_5EBF2S2KRB=GS1.1.1739567769.23.1.1739567791.0.0.0; _ga=GA1.2.1755567753.1727560849; _gid=GA1.2.1968468296.1739758667; _ga_KMRD2XMMJR=GS1.1.1739761147.21.0.1739761147.0.0.0; .DeAcFramework=202F0B84288CA98B009451432401EB9AABE6C44B6F6F0EADF05B6DACCCE176FBC14F40D7AAF2B215FE77DC1845D8B77AAA25C139346E3215A702EEDD05DAB1BA3F840FEB9A58C54E62AB72695A1DAAC7ED4B273EF6BCFE8D2D91CBC490156B0B38B4F84DAF69E4342A564D18AE64B883A4BC6AB3E75C83962CF3ACBA3E2E5670
Host: dn.rutgers.edu
Sec-Fetch-Dest: document
Sec-Fetch-Mode: navigate
Sec-Fetch-Site: none
Sec-Fetch-User: ?1
Upgrade-Insecure-Requests: 1
User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/133.0.0.0 Safari/537.36
sec-ch-ua: "Not(A:Brand";v="99", "Google Chrome";v="133", "Chromium";v="133"
sec-ch-ua-mobile: ?0
sec-ch-ua-platform: "Windows"
